+++ features/step_definitions/lib/wallets-helpers.js
@@ -14,6 +14,8 @@ export const fillOutWalletSendForm = async function(values) {
   const formSelector = '.WalletSendForm_component';
   await this.client.setValue(`${formSelector} .receiver .SimpleInput_input`, values.address);
   await this.client.setValue(`${formSelector} .amount .SimpleInput_input`, values.amount);
+  if (values.walletPassword)
+    await this.client.setValue(`${formSelector} .walletPassword .SimpleInput_input`, values.walletPassword);
   this.walletSendFormValues = values;
 };