+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changelog
 - Acceptance test for "Restore wallet with and without spending password" feature
 - Acceptance test for "Create wallet with spending password" feature
 - Acceptance test for "Import wallet with/without spending password" feature
+- Acceptance test for "Send money from a wallet with spending password" feature
 - Acceptance test for "Generate wallet address" feature
 
 ### Fixes